Add SecondIteration
In this second iteration, we are try an algorithm to manage printing the
words within the limits given from the command line and also try to
`save` a whole string into a certain memory position just by altering
the register value for a given registry and kinda traverse the string
when writing to StdOut.

Code is better modularized and we still need to figure out how to store
the string properly.

Algorithm to justify the text is in it's early stages, but looking good.

Signed-off-by: Victor Hugo M. Pinto <victor2142@gmail.com>