Fix memory leaks
Change the initial size of the hash table to be 499 to obtain better
results in the hash function, as it is a prime number, and also fix
casting.

Change Node struct to store a EntryData* and not an actual EntryData.

Fix stable_insert and stable_find function so that segmentation fault doesn't happen.

Add more verbose main function.

Signed-off-by: Victor Hugo M. Pinto <victor2142@gmail.com>