Add parse_test first iteration
Create `parse_test.c` file to implement a client program that uses the
parser() function and print its return values.

`parse_test` uses Symbol Table, implemented in PR03 and Buffer,
implemented in PR02, to read each line in a input file and to save
variables declared with IS instruction to a `alias_table`, which is
passed as an argument to parser().

Print to std_out each line parsed with it's attributes.

Signed-off-by: Victor Hugo M. Pinto <victor2142@gmail.com>